edit Auth service and edit migration

+++ EducationalInstitution.Api/Migrations/20230807091006_Table-Class-and-releted-entites.cs
@@ -161,7 +161,7 @@ namespace EducationalInstitution.Api.Migrations
                     Id = table.Column<int>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Number = table.Column<int>(type: "int", nullable: false),
-                    Capacity = table.Column<int>(type: "int", nullable: false),
+                    Capacity = table.Column<byte>(type: "tinyint", nullable: false),
                     InstructorId = table.Column<int>(type: "int", nullable: false),
                     CourseId = table.Column<int>(type: "int", nullable: false),
                     CreationDate = table.Column<DateTimeOffset>(type: "datetimeoffset", nullable: false),
